# static/swagger.json
{
  "swagger": "2.0",
  "basePath": "/api",
  "paths": {
    "/bet/client/{Username}": {
      "parameters": [
        {
          "name": "Username",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Shows a detailed list of bets from given Client username",
        "description": "*Try it out* and introduce a client username, then, hit *Execute* button to show a list of your client bets. In *Code* section you will see the list (*Response body*) and a code for a succeded or failed operation.",
        "operationId": "get_bet_list_controller",
        "tags": [
          "Bets"
        ]
      }
    },
    "/bet/{Bet_id}": {
      "parameters": [
        {
          "name": "Bet_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Shows a detailed bet from given id",
        "description": "*Try it out* and introduce a bet id you want to see; then, hit *Execute* button to show the desired bet from your database. In *Code* section you will see the bet (*Response body*) and a code for a succeded or failed operation.",
        "operationId": "get_bet_controller",
        "tags": [
          "Bets"
        ]
      },
      "delete": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Deletes a bet from given id",
        "description": "*Try it out* and introduce a bet id you want to delete; then, hit *Execute* button to delete the desired bet from your database. In *Code* section you will see the deleted bet (*Response body*) and a code for a succeded or failed operation.",
        "operationId": "delete_bet_controller",
        "tags": [
          "Bets"
        ]
      }
    },
    "/client/": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Shows a detailed list of clients",
        "description": "*Try it out* and hit *Execute* button. In *Code* section you will see a list of clients stored in your database (*Response body*) and a code for a succeded or failed operation.",
        "operationId": "get_client_list_controller",
        "tags": [
          "Clients"
        ]
      },
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Creates a new client from entry data",
        "description": "*Try it out* and introduce some values in fields below; then, hit *Execute* button to create a new client in your database. In *Code* section you will see your new client (*Response body*) and a code for a succeded or failed operation.",
        "operationId": "post_client_list_controller",
        "parameters": [
          {
            "name": "Username",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "CIF",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "E-mail",
            "in": "formData",
            "type": "string",
            "format": "email",
            "required": true
          },
          {
            "name": "Password",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "Image",
            "in": "formData",
            "type": "file"
          }
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "tags": [
          "Clients"
        ]
      }
    },
    "/client/Client/{CIF}": {
      "parameters": [
        {
          "name": "CIF",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Shows a detailed client from given CIF",
        "operationId": "get_client_controller",
        "tags": [
          "Clients"
        ]
      },
      "delete": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Deletes a client from given CIF",
        "description": "*Try it out* and introduce a client id you want to delete; then, hit *Execute* button to delete the desired client from your database. In *Code* section you will see the deleted client (*Response body*) and a code for a succeded or failed operation.",
        "operationId": "delete_client_controller",
        "tags": [
          "Clients"
        ]
      }
    },
    "/client/{Client_id}": {
      "parameters": [
        {
          "name": "Client_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "put": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Updates a client with entry data and given id",
        "description": "*Try it out* and introduce the client data and client id you want to modify; then, hit *Execute* button to apply your changes. In *Code* section you will see the modified client (*Code*) and a code for a succeded or failed operation.",
        "operationId": "put_client_controller",
        "parameters": [
          {
            "name": "Username",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "CIF",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "E-mail",
            "in": "formData",
            "type": "string",
            "format": "email"
          },
          {
            "name": "Password",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "Cash",
            "in": "formData",
            "type": "number"
          },
          {
            "name": "Image",
            "in": "formData",
            "type": "file"
          }
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "tags": [
          "Clients"
        ]
      }
    },
...
    "/user/": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Shows a detailed list of users",
        "description": "*Try it out* and hit *Execute* button. In *Code* section you will see a list of users stored in your database (*Response body*) and a code for a succeded or failed operation.",
        "operationId": "get_user_list_controller",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Creates a new user from entry data",
        "description": "*Try it out* and introduce some values in fields below; then, hit *Execute* button to create a new user in your database. In *Code* section you will see your new user (*Response body*) and a code for a succeded or failed operation.",
        "operationId": "post_user_list_controller",
        "parameters": [
          {
            "name": "Username",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "E-mail",
            "in": "formData",
            "type": "string",
            "format": "email",
            "required": true
          },
          {
            "name": "Password",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "Add Role",
            "in": "formData",
            "type": "string",
            "required": true,
            "enum": [
              "admin",
              "client"
            ]
          }
        ],
        "consumes": [
          "application/x-www-form-urlencoded",
          "multipart/form-data"
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
    "/user/{User_id}": {
      "parameters": [
        {
          "name": "User_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "put": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Updates a user with entry data and given id",
        "description": "*Try it out* and introduce the user data and user id you want to modify; then, hit *Execute* button to apply your changes. In *Code* section you will see the modified user (*Code*) and a code for a succeded or failed operation.",
        "operationId": "put_user_controller",
        "parameters": [
          {
            "name": "Username",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "E-mail",
            "in": "formData",
            "type": "string",
            "format": "email"
          },
          {
            "name": "Password",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "Add Role",
            "in": "formData",
            "type": "string",
            "enum": [
              "admin",
              "client"
            ]
          },
          {
            "name": "Delete Role",
            "in": "formData",
            "type": "string",
            "enum": [
              "admin",
              "client"
            ]
          }
        ],
        "consumes": [
          "application/x-www-form-urlencoded",
          "multipart/form-data"
        ],
        "tags": [
          "Users"
        ]
      }
    }
  },
  "info": {
    "title": "Winning Horse",
    "version": "1.0.0",
    "description": "This is the panel for admins of Winning Horse server.  For this panel, you must use the `Apikey` to authorize your credentials due to use any of methods below."
  },
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "securityDefinitions": {
    "Authentication": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Introduce your ApiKey."
    }
  },
  "security": [
    {
      "Authentication": []
    }
  ],
...
}